chore(withPropTypes): Move to addons dir
