+++ packages/forma-36-react-components/tools/.storybook/addons.js
@@ -2,4 +2,4 @@ import '@storybook/addon-storysource/register';
 import '@storybook/addon-knobs/register';
 import '@storybook/addon-notes/register';
 import '@storybook/addon-actions/register';
-import './with-props-types/register';
+import './addons/with-prop-types/register';
