+++ packages/forma-36-react-components/tools/.storybook/addons/with-prop-types/PropTypesPanel.js
