+++ packages/forma-36-react-components/tools/.storybook/config.js
@@ -3,7 +3,7 @@ import contentfulTheme from './contentful-theme';
 
 // Storybook Addon Dependencies
 import { withKnobs } from '@storybook/addon-knobs';
-import { withPropTypes } from './with-props-types/index';
+import { withPropTypes } from './addons/with-prop-types/index';
 
 // Setup Addons
 addDecorator(withKnobs);
